Fix building bug
