+++ .husky/pre-commit
@@ -1,1 +1,2 @@
 pnpm format
+pnpm build
