+++ apps/web/app/api/protected/tags/route.ts
@@ -3,7 +3,8 @@ import { NextRequest } from "next/server"
 
 export async function GET(request: NextRequest) {
   const newUrl = request.nextUrl.clone()
-  const search = newUrl.searchParams.get("search")
+  const search = newUrl.searchParams.get("search") || ""
+
   try {
     const posts = await prisma.tags.findMany({
       where: {
