+++ apps/web/package.json
@@ -50,8 +50,8 @@
     "html-react-parser": "^5.0.11",
     "lowlight": "^3.1.0",
     "lucide-react": "^0.270.0",
-    "next": "14.0.5-canary.17",
-    "next-auth": "^4.23.2",
+    "next": "14.1.0",
+    "next-auth": "^4.24.5",
     "next-themes": "^0.2.1",
     "postcss": "^8.4.28",
     "qs": "^6.11.2",
